backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiohttp==3.9.5
pydantic-settings==2.3.4
